requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
contourpy==1.3.0
cycler==0.12.1
filelock==3.13.1
fonttools==4.57.0
fsspec==2024.6.1
idna==3.4
importlib_resources==6.5.2
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.4
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
openslide-python @ file:///home/conda/feedstock_root/build_artifacts/openslide-python_1745904024913/work
packaging==25.0
pandas==2.2.3
pillow @ file:///home/conda/feedstock_root/build_artifacts/pillow_1735929703139/work
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.28.1
six==1.17.0
sympy==1.13.3
triton==2.1.0
typing_extensions==4.12.2
urllib3==1.26.13
zipp==3.21.0
tzdata==2025.1
urllib3==2.3.0
torch>=1.10.0
torchvision>=0.11.1
transformers>=4.26.0
matplotlib>=3.4.3
tqdm>=4.62.3
pycocotools>=2.0.4
timm>=0.4.12
scipy
openslide-python
openslide-bin
